services/forge_api/pyproject.toml
[project]
name = "forge-api"
version = "12.0.0"
description = "Forge Core API — REST API Server with OAuth 2.1 Resource Server"
requires-python = ">= 3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "anyio[trio]>=4.0",
    "cryptography>=42.0",
]

[tool.uv.sources]
forge-security = { path = "../../libs/forge_security", editable = true }
forge-storage = { path = "../../libs/forge_storage", editable = true }
forge-telemetry = { path = "../../libs/forge_telemetry", editable = true }
forge-a2a = { path = "../../libs/forge_a2a", editable = true }
